orchestrator_func/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-auth
google-cloud-storage